src/ai_integration.py
import os
import openai
import random
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize OpenAI API (you'll need to set OPENAI_API_KEY in your environment)
openai.api_key = os.getenv("OPENAI_API_KEY")

def analyze_video_content(video_path, target_duration, tone):
    """
    Analyze video content to find highlights (simulated implementation)
    In a real implementation, this would use computer vision and NLP APIs
    """
    # This is a placeholder implementation
    # In a real scenario, you would use:
    # 1. Video analysis APIs (Google Video Intelligence, AWS Rekognition)
    # 2. Audio transcription APIs (Google Speech-to-Text, AWS Transcribe)
    # 3. NLP APIs for summarization and highlight detection
    
    # Simulate processing by generating random highlights
    video_duration = 300  # Assume 5-minute video for simulation
    
    # Generate 3-5 random highlights based on tone
    num_highlights = random.randint(3, 5)
    highlights = []
    
    for i in range(num_highlights):
        start = random.randint(0, max(0, video_duration - target_duration))
        end = min(video_duration, start + target_duration / num_highlights)
        score = random.uniform(0.7, 1.0)  # Confidence score
        
        highlights.append({
            "start": start,
            "end": end,
            "score": score,
            "description": f"Highlight {i+1} based on {tone.lower()} tone"
        })
    
    # Sort by score and return
    highlights.sort(key=lambda x: x["score"], reverse=True)
    return highlights

def generate_teaser_clip(video_path, highlights, duration, tone, logo=None, tagline=None, add_subtitles=True, add_music=True):
    """
    Generate a teaser clip from video highlights
    """
    # This is a simplified implementation
    # In a real scenario, you would use proper video editing libraries
    
    from src.video_processor import concatenate_videoclips, add_branding, add_subtitles_to_video
    from moviepy.editor import VideoFileClip, AudioFileClip
    
    try:
        # Create output directory if it doesn't exist
        output_dir = "outputs"
        os.makedirs(output_dir, exist_ok=True)
        
        # Generate output path
        output_path = os.path.join(output_dir, f"teaser_{tone.lower()}_{int(duration)}s.mp4")
        
        # Extract highlight clips
        clips = []
        for highlight in highlights[:3]:  # Use top 3 highlights
            clip = VideoFileClip(video_path).subclip(highlight["start"], highlight["end"])
            clips.append(clip)
        
        # Concatenate clips
        if clips:
            final_clip = concatenate_videoclips(clips)
            
            # Add music if requested
            if add_music:
                # In a real implementation, you would add royalty-free music
                pass
                
            # Write the final video
            final_clip.write_videofile(output_path, codec='libx264', audio_codec='aac')
            
            # Add branding if specified
            logo_path = None
            if logo:
                # Save logo to temporary file
                with open("temp_logo.png", "wb") as f:
                    f.write(logo.getvalue())
                logo_path = "temp_logo.png"
            
            if logo_path or tagline:
                output_path = add_branding(output_path, logo_path, tagline)
            
            # Add subtitles if requested
            if add_subtitles:
                # Generate simulated subtitles
                subtitles = [
                    (0.0, 5.0, "Discover something amazing"),
                    (5.0, 10.0, "With our innovative solution"),
                    (10.0, 15.0, "Join us today!")
                ]
                output_path = add_subtitles_to_video(output_path, subtitles, output_path)
            
            # Clean up temporary files
            if logo_path and os.path.exists(logo_path):
                os.remove(logo_path)
                
            return output_path
        
    except Exception as e:
        print(f"Error generating teaser: {e}")
        raise Exception("Failed to generate teaser clip")

def generate_caption(tone="Professional"):
    """
    Generate a social media caption using AI
    """
    # This is a simplified implementation
    # In a real scenario, you would use the OpenAI API or similar
    
    try:
        # Simulate AI-generated caption based on tone
        captions = {
            "Professional": "Introducing our latest solution designed to enhance productivity and efficiency. #Innovation #Tech",
            "Exciting": "🎉 Get ready for something amazing! Our new release will transform how you work. #GameChanger #Excited",
            "Educational": "Learn how our new approach can help solve common challenges in your industry. #Education #Knowledge",
            "Inspirational": "Unlock your potential with tools designed to help you achieve more. #Inspiration #Growth"
        }
        
        return captions.get(tone, "Check out our latest content!")
        
    except Exception as e:
        print(f"Error generating caption: {e}")
        return "Check out this teaser for our latest content!"