requirements.txt
streamlit==1.29.0
openai==1.3.0
moviepy==1.0.3
pytube==15.0.0
requests==2.31.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydub==0.25.1
streamlit-player==0.1.6
python-magic==0.4.27
pillow==10.0.1